[DM-23942] Restructure some session code
Start moving code for managing oauth2_proxy sessions and their
corresponding tickets into separate modules.  Move add_padding
into a utility module since it's used from multiple places.

Rename TokenStore to SessionStore to more accurately reflect what
it does.
