+++ docs/api.rst
@@ -14,5 +14,9 @@ API reference
 .. automodapi:: jwt_authorizer.config
    :no-inherited-members:
 
+.. automodapi:: jwt_authorizer.session
+
 .. automodapi:: jwt_authorizer.tokens
    :no-inherited-members:
+
+.. automodapi:: jwt_authorizer.util
