+++ src/jwt_authorizer/tokens.py
@@ -8,13 +8,10 @@ import hmac
 import json
 import logging
 import os
-import re
 import struct
-from binascii import Error
 from calendar import timegm
 from dataclasses import dataclass
-from dataclasses import field as dc_field
-from datetime import datetime, timedelta, timezone
+from datetime import datetime, timedelta
 from typing import TYPE_CHECKING, cast
 
 import jwt
@@ -30,22 +27,19 @@ from flask_wtf import FlaskForm
 from wtforms import BooleanField, HiddenField, SubmitField
 
 from jwt_authorizer import config
+from jwt_authorizer.session import Ticket
+from jwt_authorizer.util import add_padding
 
 __all__ = [
     "AlterTokenForm",
     "Issuer",
-    "Session",
-    "Ticket",
-    "TokenStore",
     "TokenVerifier",
     "add_padding",
     "api_capabilities_token_form",
-    "create_token_store",
     "create_token_verifier",
     "get_key_as_pem",
     "get_tokens",
     "issue_token",
-    "parse_ticket",
     "revoke_token",
 ]
 
@@ -55,82 +49,6 @@ if TYPE_CHECKING:
     from typing import Any, Dict, List, Mapping, Optional, Tuple
 
 logger = logging.getLogger(__name__)
-
-
-def new_id() -> str:
-    """Generate a new ticket ID."""
-    return hashlib.sha1(os.urandom(16)).hexdigest()
-
-
-def new_secret() -> bytes:
-    """Generate a new ticket encryption secret."""
-    return os.urandom(16)
-
-
-@dataclass
-class Ticket:
-    """A class represeting an oauth2_proxy ticket."""
-
-    ticket_id: str = dc_field(default_factory=new_id)
-    secret: bytes = dc_field(default_factory=new_secret)
-
-    def as_handle(self, prefix: str) -> str:
-        """Return the handle for this ticket.
-
-        Parameters
-        ----------
-        prefix : `str`
-            Prefix to prepend to the ticket ID.
-        """
-        return f"{prefix}-{self.ticket_id}"
-
-    def encode(self, prefix: str) -> str:
-        """Return the encoded ticket, suitable for putting in a cookie.
-
-        Parameters
-        ----------
-        prefix : `str`
-            Prefix to prepend to the ticket ID.
-        """
-        secret_b64 = base64.urlsafe_b64encode(self.secret).decode().rstrip("=")
-        return f"{prefix}-{self.ticket_id}.{secret_b64}"
-
-
-def parse_ticket(prefix: str, ticket: str) -> Optional[Ticket]:
-    """Parse an oauth2_proxy ticket string into a Ticket.
-
-    Parameters
-    ----------
-    prefix : `str`
-        The expected prefix for the ticket.
-    ticket : `str`
-        The encoded ticket string.
-
-    Returns
-    -------
-    decoded_ticket : Optional[`Ticket`]
-        The decoded Ticket, or None if there was an error.
-    """
-    full_prefix = f"{prefix}-"
-    if not ticket.startswith(full_prefix):
-        logger.error("Error decoding ticket: Ticket not in expected format")
-        return None
-    trimmed_ticket = ticket[len(full_prefix) :]
-    if "." not in trimmed_ticket:
-        logger.error("Error decoding ticket: Ticket not in expected format")
-        return None
-    ticket_id, secret_b64 = trimmed_ticket.split(".")
-    try:
-        int(ticket_id, 16)  # Check hex
-        secret = base64.b64decode(
-            add_padding(secret_b64), altchars=b"-_", validate=True
-        )
-        if secret == b"":
-            raise ValueError("ticket secret is empty")
-        return Ticket(ticket_id=ticket_id, secret=secret)
-    except (ValueError, Error) as e:
-        logger.error("Error decoding ticket: %s", str(e))
-        return None
 
 
 def issue_token(
@@ -642,26 +560,6 @@ def _o2proxy_signed_session(session_payload: str) -> str:
     )
 
 
-def add_padding(encoded: str) -> str:
-    """Add padding to base64 encoded bytes.
-
-    Parameters
-    ----------
-    encoded : `str`
-        A base64-encoded string, possibly with the padding removed.
-
-    Returns
-    -------
-    result : `str`
-        A correctly-padded version of the encoded string.
-    """
-    underflow = len(encoded) % 4
-    if underflow:
-        return encoded + ("=" * (4 - underflow))
-    else:
-        return encoded
-
-
 def user_tokens_redis_key(user_id: str) -> str:
     """The Redis key for storing a user's tokens.
 
@@ -676,180 +574,6 @@ def user_tokens_redis_key(user_id: str) -> str:
         The Redis key under which that user's tokens will be stored.
     """
     return f"tokens:{user_id}"
-
-
-@dataclass
-class Session:
-    """An oauth2_proxy session.
-
-    Tokens are currently stored in Redis as a JSON dump of a dictionary.  This
-    class represents the deserialized form of a session.
-    """
-
-    token: str
-    email: str
-    user: str
-    created_at: datetime
-    expires_on: datetime
-
-
-class TokenStore:
-    """Stores tokens and retrieves them by ticket.
-
-    Parameters
-    ----------
-    prefix : `str`
-        Prefix used for storing oauth2_proxy session state.
-    redis : `redis.Redis`
-        A Redis client configured to talk to the backend store that holds the
-        (encrypted) tokens.
-    key : `bytes`
-        Encryption key for the individual components of the stored session.
-    """
-
-    def __init__(self, prefix: str, key: bytes, redis: redis.Redis) -> None:
-        self.prefix = prefix
-        self.key = key
-        self.redis = redis
-
-    def get_session(self, ticket: Ticket) -> Optional[Session]:
-        """Retrieve and decrypt the session for a ticket.
-
-        Parameters
-        ----------
-        ticket : `Ticket`
-            The ticket corresponding to the token.
-
-        Returns
-        -------
-        session : `Session` or `None`
-            The corresponding session, or `None` if no session exists for this
-            ticket.
-        """
-        handle = ticket.as_handle(self.prefix)
-        encrypted_session = self.redis.get(handle)
-        if not encrypted_session:
-            return None
-
-        return self._decrypt_session(ticket.secret, encrypted_session)
-
-    def _decrypt_session(
-        self, secret: bytes, encrypted_session: bytes
-    ) -> Session:
-        """Decrypt an oauth2_proxy session.
-
-        Parameters
-        ----------
-        secret : `bytes`
-            Decryption key.
-        encrypted_session : `bytes`
-            The encrypted session.
-
-        Returns
-        -------
-        session : `Sesssion`
-            The decrypted sesssion.
-        """
-        cipher = Cipher(
-            algorithms.AES(secret), modes.CFB(secret), default_backend()
-        )
-        decryptor = cipher.decryptor()
-        session_dict = json.loads(
-            decryptor.update(encrypted_session) + decryptor.finalize()
-        )
-        return Session(
-            token=self._decrypt_session_component(session_dict["IDToken"]),
-            email=self._decrypt_session_component(session_dict["Email"]),
-            user=self._decrypt_session_component(session_dict["User"]),
-            created_at=self._parse_session_date(session_dict["CreatedAt"]),
-            expires_on=self._parse_session_date(session_dict["ExpiresOn"]),
-        )
-
-    def _decrypt_session_component(self, encrypted_str: str) -> str:
-        """Decrypt a component of an encrypted oauth2_proxy session.
-
-        Parameters
-        ----------
-        encrypted_str : `str`
-            The encrypted field with its IV prepended.
-
-        Returns
-        -------
-        component : `str`
-            The decrypted value.
-        """
-        encrypted_bytes = base64.b64decode(encrypted_str)
-        iv = encrypted_bytes[:16]
-        cipher = Cipher(
-            algorithms.AES(self.key), modes.CFB(iv), default_backend()
-        )
-        decryptor = cipher.decryptor()
-        field = decryptor.update(encrypted_bytes[16:]) + decryptor.finalize()
-        return field.decode()
-
-    @staticmethod
-    def _parse_session_date(date_str: str) -> datetime:
-        """Parse a date from a session record.
-
-        Parameters
-        ----------
-        date_str : `str`
-            The date in string format.
-
-        Returns
-        -------
-        date : `datetime`
-            The parsed date.
-
-        Notes
-        -----
-        This date may be written by oauth2_proxy instead of us, in which case
-        it will use a Go date format that includes fractional seconds down to
-        the nanosecond.  Python doesn't have a date format that parses this,
-        so the fractional seconds portion will be dropped, leading to an
-        inaccuracy of up to a second.
-        """
-        date_str = re.sub("[.][0-9]+Z$", "Z", date_str)
-        date = datetime.strptime(date_str, "%Y-%m-%dT%H:%M:%SZ")
-        return date.replace(tzinfo=timezone.utc)
-
-
-def get_redis_client(app: Flask) -> redis.Redis:
-    """Get a Redis client from the Flask application pool.
-
-    Exists primarily to be overridden by tests.
-
-    Parameters
-    ----------
-    app : `flask.Flask`
-        The Flask application.
-
-    Returns
-    -------
-    redis_client : `redis.Redis`
-        A Redis client.
-    """
-    return redis.Redis(connection_pool=app.redis_pool)
-
-
-def create_token_store(app: Flask) -> TokenStore:
-    """Create a TokenStore from a Flask app configuration.
-
-    Parameters
-    ----------
-    app : `flask.Flask`
-        The Flask application.
-
-    Returns
-    -------
-    token_store : `TokenStore`
-        A TokenStore created from that Flask application configuration.
-    """
-    redis_client = get_redis_client(app)
-    prefix = app.config["OAUTH2_STORE_SESSION"]["TICKET_PREFIX"]
-    secret_str = app.config["OAUTH2_STORE_SESSION"]["OAUTH2_PROXY_SECRET"]
-    secret = base64.urlsafe_b64decode(secret_str)
-    return TokenStore(prefix, secret, redis_client)
 
 
 @dataclass
